// models/course.ts
export type course = {
  course_name: string;
  createdAt: string;
  description: string;
  duration: string;
  enrollmentStatus: string;
  id: string;
  instructor_name: string;
  location: string;
  prerequisites: string[];
  price: number;
  schedule: string;
  thumbnail: string;
  updatedAt: string;
};